prevent possible overflow

+++ include/meta.h
@@ -24,12 +24,12 @@ namespace meta
     /**
      * Numbering system for string terms.
      */
-    typedef uint16_t term_id;
+    typedef uint32_t term_id;
 
     /**
      * Numbering system for Documents.
      */
-    typedef uint16_t doc_id;
+    typedef uint32_t doc_id;
 
     /**
      * Algorithms for feature selection, KNN search, and confusion
